No more absolute paths and 32-bit fixes
+ Do not include netcomm.mk with an "pseudo-absolute" path. After fixing
. the qa target this can be relative.
+ A few 32-bit system fixes in QA apps.

Tim


git-svn-id: https://lechuck.informatik.rwth-aachen.de:2342/svn/fawkes/trunk@231 fc82b16a-671c-0410-8732-9903a8d2dd0d

+++ src/libs/netcomm/qa/qa_dynamic_buffer.cpp
@@ -61,7 +61,7 @@ main(int argc, char **argv)
     size_t size;
     void *buf = dr->next(&size);
     strncpy(tmp, (const char *)buf, size);
-    printf("Read string (%lu bytes): '%s'\n", size, tmp);
+    printf("Read string (%lu bytes): '%s'\n", (unsigned long int)size, tmp);
   }
 
   delete dw;
